+++ src/globals.hpp
@@ -43,5 +43,5 @@ extern Chess * chess;
 extern LightModel * lights;
 extern GraphicModelChess * chessTable, * selectedFrame;
 extern vector<GraphicModelChess> previewPositions, pieceModels;
-extern Point3D<float> tableRotation;
+extern Point3D<int> tableRotation;
 #endif
